requirements.txt
opencv-python~=4.5.5.64
imutils~=0.5.4
argparse~=1.4.0
Pillow~=9.1.0
tensorflow~=2.8.0
tensorflow-gpu~=2.8.0
matplotlib~=3.5.1
numpy~=1.22.3
keras~=2.8.0
scikit-learn~=1.0.2